Create separate styles files for Home and CardRow components
