+++ LabProject/src/components/home/AccountsOverview.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
-import { StyleSheet, Text, View } from 'react-native';
-import theme from '../../theme/theme';
+import { Text, View } from 'react-native';
 import { getTotalPrice } from '../../utils';
-import CardRow, { ICardRowProps } from '../common/CardRow';
+import CardRow, { ICardRowProps } from '../common/card-row/CardRow';
 import Routes from '../../constants/routes';
+import { accountsOverviewStyles as styles } from './styles';
 
 export interface IAccountsOverviewProps {
     onHandleNavigation: (route: keyof typeof Routes) => void;
@@ -41,79 +41,5 @@ const AccountsOverview = ({ accountOverviewScreens, onHandleNavigation }: IAccou
         </View>
     );
 };
-
-const styles = StyleSheet.create({
-    root: {
-        alignSelf: 'stretch',
-        display: 'flex',
-        alignItems: 'center',
-        flexDirection: 'column',
-        paddingTop: 15,
-        backgroundColor: theme.colors?.white,
-        borderRadius: 5,
-        borderWidth: 1,
-        borderColor: `rgba(${theme.colors.grey1Rgba},0.3)`,
-    },
-    mainHeaderTitle: {
-        fontSize: 22,
-        fontWeight: '600',
-    },
-    priceTitle: {
-        fontSize: 30,
-        marginTop: 10,
-        marginBottom: 5,
-    },
-    priceHelperTitle: {
-        color: theme.colors?.grey0,
-        textTransform: 'capitalize',
-    },
-    cardRowContainer: {
-        width: '100%',
-        display: 'flex',
-        flexDirection: 'row',
-        justifyContent: 'space-between',
-        paddingHorizontal: 20,
-        paddingVertical: 10,
-    },
-    cardRowWithBorder: {
-        borderBottomWidth: 1,
-        borderBottomColor: `rgba(${theme.colors.grey1Rgba},0.3)`,
-    },
-    cardRowWithoutBorder: {
-        borderBottomWidth: undefined,
-    },
-    cardRowLeft: {
-        display: 'flex',
-        flexDirection: 'column',
-    },
-    leftTitle: {
-        fontSize: 18,
-    },
-    letTitleIcon: {
-        width: 15,
-        height: 15,
-        resizeMode: 'contain',
-    },
-    leftSubtitle: {
-        marginTop: 5,
-        color: theme.colors?.grey0,
-        fontSize: 14,
-    },
-    cardRowRight: {
-        display: 'flex',
-        flexDirection: 'row',
-        alignItems: 'center',
-    },
-    rightTitle: {
-        fontSize: 24,
-        marginRight: 10,
-    },
-    rightIcon: {
-        width: 15,
-        height: 15,
-        tintColor: theme.colors?.primary,
-        transform: [{ rotate: '180deg' }],
-    },
-});
 
 export default AccountsOverview;
