+++ LabProject/src/components/home/GoodnessCard.tsx
@@ -1,9 +1,9 @@
 import React, { useState } from 'react';
-import { Image, ImageSourcePropType, StyleSheet, Text, TouchableOpacity, View } from 'react-native';
-import theme from '../../theme/theme';
+import { Image, ImageSourcePropType, Text, TouchableOpacity, View } from 'react-native';
 import { Badge } from 'react-native-elements';
 import Button, { ButtonColor } from '../common/button/Button';
 import Video from 'react-native-video';
+import { goodnessCardStyles as styles } from './styles';
 
 export interface IGoodnessCardProps {
     avatarLink: ImageSourcePropType;
@@ -93,86 +93,5 @@ const GoodnessCard: React.FC<IGoodnessCardProps> = ({
         </View>
     );
 };
-
-const styles = StyleSheet.create({
-    root: {
-        flex: 1,
-        alignSelf: 'stretch',
-        backgroundColor: theme.colors?.white,
-        borderRadius: 5,
-        borderWidth: 1,
-        borderColor: `rgba(${theme.colors.grey1Rgba},0.3)`,
-    },
-    header: {
-        display: 'flex',
-        flexDirection: 'row',
-        alignItems: 'center',
-        padding: 10,
-    },
-    headerTitlesSummaryContainer: {
-        display: 'flex',
-        flexDirection: 'column',
-        justifyContent: 'space-around',
-        marginLeft: 15,
-        height: 50,
-    },
-    headerTitleContainer: {
-        display: 'flex',
-        flexDirection: 'row',
-        alignItems: 'center',
-    },
-    title: {
-        fontSize: 19,
-        color: theme.colors.black,
-    },
-    subtitle: {
-        fontSize: 15,
-        color: theme.colors.grey0,
-    },
-    badge: {
-        backgroundColor: theme.colors.primary,
-        marginHorizontal: 5,
-    },
-    avatarIcon: {
-        width: 50,
-        height: 50,
-    },
-    image: {
-        width: undefined,
-        height: 175,
-        resizeMode: 'cover',
-    },
-    body: {
-        paddingHorizontal: 10,
-        paddingVertical: 20,
-    },
-    mainText: {
-        marginBottom: 20,
-        fontSize: 15,
-        color: theme.colors.black,
-    },
-    buttonIcon: {
-        width: 14,
-        height: 14,
-        marginRight: 10,
-    },
-    videoContainer: {
-        position: 'relative',
-    },
-    video: {
-        width: undefined,
-        height: 200,
-    },
-    videoSoundIconContainer: {
-        position: 'absolute',
-        bottom: 5,
-        right: 5,
-    },
-    videoSoundIcon: {
-        height: 30,
-        width: 30,
-        resizeMode: 'contain',
-    },
-});
 
 export default GoodnessCard;
